[REM-103] feat: implement server management features
- server/__main__.py: added main module to run the server with Uvicorn
- server/templates/index.html: created basic HTML template for server responses
- .vscode/launch.json: added configuration for Application Manager debugging
- server/blueprints/asgi/__init__.py: enhanced ASGI server management with process handling
- server/blueprints/celery_worker/__init__.py: improved worker server management with process handling
- server/blueprints/celery_beat/__init__.py: added Celery Beat server management with process handling
- server/blueprints/celery_beat/celery_beat.py: removed obsolete Celery Beat blueprint
