+++ .vscode/launch.json
@@ -82,6 +82,12 @@
       "type": "debugpy",
       "request": "launch",
       "module": "app.beat"
+    },
+    {
+      "name": "Application Manager",
+      "type": "debugpy",
+      "request": "launch",
+      "module": "server"
     }
   ]
 }
