+++ server/__main__.py
@@ -1,0 +1,17 @@
+"""Main module for the server."""
+
+import asyncio
+
+import uvicorn
+
+from . import create_app
+
+if __name__ == "__main__":
+    app = asyncio.run(create_app())
+    hostname = "127.0.0.1"
+    port = 5002
+    uvicorn.run(
+        app,
+        host=hostname,
+        port=port,
+    )
