+++ server/templates/index.html
@@ -1,0 +1,16 @@
+<!DOCTYPE html>
+<html lang="en">
+
+  <head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>Document</title>
+  </head>
+
+  <body>
+    {% if message %}
+    <h1>{{ message }}</h1>
+    {% endif %}
+  </body>
+
+</html>
